Some basic cleanup in primitives/account.
