+++ primitives/account/tests/htlc_contract.rs
@@ -1,11 +1,10 @@
 use std::convert::TryInto;
 
 use beserial::{Deserialize, Serialize, SerializingError};
-use nimiq_account::{
-    AccountError, AccountTransactionInteraction, AccountType, HashedTimeLockedContract,
-};
+use nimiq_account::{AccountError, AccountTransactionInteraction, HashedTimeLockedContract};
 use nimiq_hash::{Blake2bHasher, HashOutput, Hasher, Sha256Hasher};
 use nimiq_keys::{Address, KeyPair, PrivateKey};
+use nimiq_primitives::account::AccountType;
 use nimiq_primitives::coin::Coin;
 use nimiq_primitives::networks::NetworkId;
 use nimiq_transaction::account::htlc_contract::{AnyHash, HashAlgorithm, ProofType};
